WIP: reorganize io-module
